pyproject.toml
[project]
name = "ai-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.182.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.4",
    "setuptools>=80.9.0",
    "streamlit>=1.49.1",
]
